python/requirements.txt
openai==1.12.0
PyMuPDF==1.23.21
python-docx==1.1.0
nltk==3.8.1
python-dotenv==1.0.1
requests==2.31.0